Fix and improve `providers` package
