+++ packages/providers/package.json
@@ -10,7 +10,7 @@
     "build": "tsc -b tsconfig.build.json",
     "test": "mocha -r ts-node/register 'src/**/**.test.ts'",
     "test:coverage": "nyc --reporter=text mocha -r ts-node/register 'src/**/**.test.ts'",
-    "generate-operation-types": "graphql-codegen"
+    "generate-operations": "graphql-codegen"
   },
   "license": "UNLICENSED",
   "dependencies": {
@@ -20,12 +20,14 @@
     "@fuel-ts/abi-coder": "^0.0.1",
     "@fuel-ts/merkle": "^0.0.1",
     "@fuel-ts/transactions": "^0.0.1",
-    "cross-fetch": "^3.1.4",
-    "graphql": "^15.6.1"
+    "graphql": "^15.6.1",
+    "graphql-request": "^3.6.1",
+    "graphql-tag": "^2.12.6"
   },
   "devDependencies": {
     "@graphql-codegen/cli": "^2.2.1",
     "@graphql-codegen/typescript": "^2.2.4",
+    "@graphql-codegen/typescript-graphql-request": "^4.3.0",
     "@graphql-codegen/typescript-operations": "^2.1.8",
     "chai": "^4.3.4"
   }
