--- packages/providers/src/operations.types.ts
@@ -1,460 +1,0 @@
-export type Maybe<T> = T | null;
-export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
-export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
-export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
-/** All built-in and custom scalars, mapped to their actual values */
-export type Scalars = {
-  ID: string;
-  String: string;
-  Boolean: boolean;
-  Int: number;
-  Float: number;
-  /**
-   * Implement the DateTime<Utc> scalar
-   *
-   * The input/output is a string in RFC3339 format.
-   */
-  DateTime: any;
-  HexString: any;
-  HexString256: any;
-};
-
-export type Block = {
-  __typename?: 'Block';
-  height: Scalars['Int'];
-  id: Scalars['HexString256'];
-  producer: Scalars['HexString256'];
-  time: Scalars['DateTime'];
-  transactions: Array<Transaction>;
-};
-
-export type BlockConnection = {
-  __typename?: 'BlockConnection';
-  /** A list of edges. */
-  edges?: Maybe<Array<Maybe<BlockEdge>>>;
-  /** Information to aid in pagination. */
-  pageInfo: PageInfo;
-};
-
-/** An edge in a connection. */
-export type BlockEdge = {
-  __typename?: 'BlockEdge';
-  /** A cursor for use in pagination */
-  cursor: Scalars['String'];
-  /** The item at the end of the edge */
-  node: Block;
-};
-
-export type ChainInfo = {
-  __typename?: 'ChainInfo';
-  baseChainHeight: Scalars['Int'];
-  latestBlock: Block;
-  name: Scalars['String'];
-  peerCount: Scalars['Int'];
-};
-
-export type ChangeOutput = {
-  __typename?: 'ChangeOutput';
-  amount: Scalars['Int'];
-  color: Scalars['HexString256'];
-  to: Scalars['HexString256'];
-};
-
-export type Coin = {
-  __typename?: 'Coin';
-  amount: Scalars['Int'];
-  blockCreated: Scalars['Int'];
-  color: Scalars['HexString256'];
-  id: Scalars['HexString256'];
-  maturity: Scalars['Int'];
-  owner: Scalars['HexString256'];
-  status: CoinStatus;
-};
-
-export type CoinConnection = {
-  __typename?: 'CoinConnection';
-  /** A list of edges. */
-  edges?: Maybe<Array<Maybe<CoinEdge>>>;
-  /** Information to aid in pagination. */
-  pageInfo: PageInfo;
-};
-
-/** An edge in a connection. */
-export type CoinEdge = {
-  __typename?: 'CoinEdge';
-  /** A cursor for use in pagination */
-  cursor: Scalars['String'];
-  /** The item at the end of the edge */
-  node: Coin;
-};
-
-export type CoinOutput = {
-  __typename?: 'CoinOutput';
-  amount: Scalars['Int'];
-  color: Scalars['HexString256'];
-  to: Scalars['HexString256'];
-};
-
-export enum CoinStatus {
-  Spent = 'SPENT',
-  Unspent = 'UNSPENT'
-}
-
-export type ContractCreated = {
-  __typename?: 'ContractCreated';
-  contractId: Scalars['HexString256'];
-};
-
-export type ContractOutput = {
-  __typename?: 'ContractOutput';
-  balanceRoot: Scalars['HexString256'];
-  inputIndex: Scalars['Int'];
-  stateRoot: Scalars['HexString256'];
-};
-
-export type FailureStatus = {
-  __typename?: 'FailureStatus';
-  blockId: Scalars['HexString256'];
-  reason: Scalars['String'];
-  time: Scalars['DateTime'];
-};
-
-export type Input = InputCoin | InputContract;
-
-export type InputCoin = {
-  __typename?: 'InputCoin';
-  amount: Scalars['Int'];
-  color: Scalars['HexString256'];
-  maturity: Scalars['Int'];
-  owner: Scalars['HexString256'];
-  predicate: Scalars['HexString'];
-  predicateData: Scalars['HexString'];
-  utxoId: Scalars['HexString256'];
-  witnessIndex: Scalars['Int'];
-};
-
-export type InputContract = {
-  __typename?: 'InputContract';
-  balanceRoot: Scalars['HexString256'];
-  contractId: Scalars['HexString256'];
-  stateRoot: Scalars['HexString256'];
-  utxoId: Scalars['HexString256'];
-};
-
-export type Mutation = {
-  __typename?: 'Mutation';
-  /** dry-run the transaction using a fork of current state, no changes are committed. */
-  dryRun: Array<Receipt>;
-  endSession: Scalars['Boolean'];
-  execute: Scalars['Boolean'];
-  reset: Scalars['Boolean'];
-  startSession: Scalars['ID'];
-  /** Submits transaction to the pool */
-  submit: Scalars['HexString256'];
-};
-
-
-export type MutationDryRunArgs = {
-  tx: Scalars['HexString'];
-};
-
-
-export type MutationEndSessionArgs = {
-  id: Scalars['ID'];
-};
-
-
-export type MutationExecuteArgs = {
-  id: Scalars['ID'];
-  op: Scalars['String'];
-};
-
-
-export type MutationResetArgs = {
-  id: Scalars['ID'];
-};
-
-
-export type MutationSubmitArgs = {
-  tx: Scalars['HexString'];
-};
-
-export type Output = ChangeOutput | CoinOutput | ContractCreated | ContractOutput | VariableOutput | WithdrawalOutput;
-
-/** Information about pagination in a connection */
-export type PageInfo = {
-  __typename?: 'PageInfo';
-  /** When paginating forwards, the cursor to continue. */
-  endCursor?: Maybe<Scalars['String']>;
-  /** When paginating forwards, are there more items? */
-  hasNextPage: Scalars['Boolean'];
-  /** When paginating backwards, are there more items? */
-  hasPreviousPage: Scalars['Boolean'];
-  /** When paginating backwards, the cursor to continue. */
-  startCursor?: Maybe<Scalars['String']>;
-};
-
-export type Query = {
-  __typename?: 'Query';
-  block?: Maybe<Block>;
-  blocks: BlockConnection;
-  chain: ChainInfo;
-  coin?: Maybe<Coin>;
-  coinsByOwner: CoinConnection;
-  /** Returns true when the GraphQL API is serving requests. */
-  health: Scalars['Boolean'];
-  memory: Scalars['String'];
-  register: Scalars['Int'];
-  transaction?: Maybe<Transaction>;
-  transactions: TransactionConnection;
-  version: Scalars['String'];
-};
-
-
-export type QueryBlockArgs = {
-  height?: Maybe<Scalars['Int']>;
-  id?: Maybe<Scalars['HexString256']>;
-};
-
-
-export type QueryBlocksArgs = {
-  after?: Maybe<Scalars['String']>;
-  before?: Maybe<Scalars['String']>;
-  first?: Maybe<Scalars['Int']>;
-  last?: Maybe<Scalars['Int']>;
-};
-
-
-export type QueryCoinArgs = {
-  id: Scalars['HexString256'];
-};
-
-
-export type QueryCoinsByOwnerArgs = {
-  after?: Maybe<Scalars['String']>;
-  before?: Maybe<Scalars['String']>;
-  first?: Maybe<Scalars['Int']>;
-  last?: Maybe<Scalars['Int']>;
-  owner: Scalars['HexString256'];
-};
-
-
-export type QueryMemoryArgs = {
-  id: Scalars['ID'];
-  size: Scalars['Int'];
-  start: Scalars['Int'];
-};
-
-
-export type QueryRegisterArgs = {
-  id: Scalars['ID'];
-  register: Scalars['Int'];
-};
-
-
-export type QueryTransactionArgs = {
-  id: Scalars['HexString256'];
-};
-
-
-export type QueryTransactionsArgs = {
-  after?: Maybe<Scalars['String']>;
-  before?: Maybe<Scalars['String']>;
-  first?: Maybe<Scalars['Int']>;
-  last?: Maybe<Scalars['Int']>;
-};
-
-export type Receipt = {
-  __typename?: 'Receipt';
-  a?: Maybe<Scalars['Int']>;
-  amount?: Maybe<Scalars['Int']>;
-  b?: Maybe<Scalars['Int']>;
-  color?: Maybe<Scalars['HexString256']>;
-  digest?: Maybe<Scalars['HexString256']>;
-  gas?: Maybe<Scalars['Int']>;
-  id: Scalars['HexString256'];
-  is: Scalars['Int'];
-  len?: Maybe<Scalars['Int']>;
-  pc: Scalars['Int'];
-  ptr?: Maybe<Scalars['Int']>;
-  ra?: Maybe<Scalars['Int']>;
-  rawPayload: Scalars['HexString'];
-  rb?: Maybe<Scalars['Int']>;
-  rc?: Maybe<Scalars['Int']>;
-  rd?: Maybe<Scalars['Int']>;
-  reason?: Maybe<Scalars['Int']>;
-  receiptType: ReceiptType;
-  to?: Maybe<Scalars['HexString256']>;
-  toAddress?: Maybe<Scalars['HexString256']>;
-  val?: Maybe<Scalars['Int']>;
-};
-
-export enum ReceiptType {
-  Call = 'CALL',
-  Log = 'LOG',
-  LogData = 'LOG_DATA',
-  Panic = 'PANIC',
-  Return = 'RETURN',
-  ReturnData = 'RETURN_DATA',
-  Revert = 'REVERT',
-  Transfer = 'TRANSFER',
-  TransferOut = 'TRANSFER_OUT'
-}
-
-export type SubmittedStatus = {
-  __typename?: 'SubmittedStatus';
-  time: Scalars['DateTime'];
-};
-
-export type SuccessStatus = {
-  __typename?: 'SuccessStatus';
-  blockId: Scalars['HexString256'];
-  programState: Scalars['HexString'];
-  time: Scalars['DateTime'];
-};
-
-export type Transaction = {
-  __typename?: 'Transaction';
-  bytecodeWitnessIndex?: Maybe<Scalars['Int']>;
-  gasLimit: Scalars['Int'];
-  gasPrice: Scalars['Int'];
-  id: Scalars['HexString256'];
-  inputColors: Array<Scalars['HexString256']>;
-  inputContracts: Array<Scalars['HexString256']>;
-  inputs: Array<Input>;
-  isScript: Scalars['Boolean'];
-  maturity: Scalars['Int'];
-  outputs: Array<Output>;
-  /** Return the transaction bytes using canonical encoding */
-  rawPayload: Scalars['HexString'];
-  receipts?: Maybe<Array<Receipt>>;
-  receiptsRoot?: Maybe<Scalars['HexString256']>;
-  salt?: Maybe<Scalars['HexString256']>;
-  script?: Maybe<Scalars['HexString']>;
-  scriptData?: Maybe<Scalars['HexString']>;
-  staticContracts?: Maybe<Array<Scalars['HexString256']>>;
-  status?: Maybe<TransactionStatus>;
-  witnesses: Array<Scalars['HexString']>;
-};
-
-export type TransactionConnection = {
-  __typename?: 'TransactionConnection';
-  /** A list of edges. */
-  edges?: Maybe<Array<Maybe<TransactionEdge>>>;
-  /** Information to aid in pagination. */
-  pageInfo: PageInfo;
-};
-
-/** An edge in a connection. */
-export type TransactionEdge = {
-  __typename?: 'TransactionEdge';
-  /** A cursor for use in pagination */
-  cursor: Scalars['String'];
-  /** The item at the end of the edge */
-  node: Transaction;
-};
-
-export type TransactionStatus = FailureStatus | SubmittedStatus | SuccessStatus;
-
-export type VariableOutput = {
-  __typename?: 'VariableOutput';
-  amount: Scalars['Int'];
-  color: Scalars['HexString256'];
-  to: Scalars['HexString256'];
-};
-
-export type WithdrawalOutput = {
-  __typename?: 'WithdrawalOutput';
-  amount: Scalars['Int'];
-  color: Scalars['HexString256'];
-  to: Scalars['HexString256'];
-};
-
-export type GetVersionQueryVariables = Exact<{ [key: string]: never; }>;
-
-
-export type GetVersionQuery = { __typename?: 'Query', version: string };
-
-export type GetTransactionQueryVariables = Exact<{
-  transactionId: Scalars['HexString256'];
-}>;
-
-
-export type GetTransactionQuery = { __typename?: 'Query', transaction?: { __typename?: 'Transaction', id: any, rawPayload: any } | null | undefined };
-
-export type GetTransactionsQueryVariables = Exact<{
-  after?: Maybe<Scalars['String']>;
-  before?: Maybe<Scalars['String']>;
-  first?: Maybe<Scalars['Int']>;
-  last?: Maybe<Scalars['Int']>;
-}>;
-
-
-export type GetTransactionsQuery = { __typename?: 'Query', transactions: { __typename?: 'TransactionConnection', edges?: Array<{ __typename?: 'TransactionEdge', node: { __typename?: 'Transaction', id: any, rawPayload: any } } | null | undefined> | null | undefined } };
-
-export type BlockFragmentFragment = { __typename?: 'Block', id: any, height: number, producer: any, time: any, transactions: Array<{ __typename?: 'Transaction', id: any, rawPayload: any }> };
-
-export type GetBlockQueryVariables = Exact<{
-  blockId: Scalars['HexString256'];
-}>;
-
-
-export type GetBlockQuery = { __typename?: 'Query', block?: { __typename?: 'Block', id: any, height: number, producer: any, time: any, transactions: Array<{ __typename?: 'Transaction', id: any, rawPayload: any }> } | null | undefined };
-
-export type GetBlocksQueryVariables = Exact<{
-  after?: Maybe<Scalars['String']>;
-  before?: Maybe<Scalars['String']>;
-  first?: Maybe<Scalars['Int']>;
-  last?: Maybe<Scalars['Int']>;
-}>;
-
-
-export type GetBlocksQuery = { __typename?: 'Query', blocks: { __typename?: 'BlockConnection', edges?: Array<{ __typename?: 'BlockEdge', node: { __typename?: 'Block', id: any, height: number, producer: any, time: any, transactions: Array<{ __typename?: 'Transaction', id: any, rawPayload: any }> } } | null | undefined> | null | undefined } };
-
-export type GetCoinQueryVariables = Exact<{
-  coinId: Scalars['HexString256'];
-}>;
-
-
-export type GetCoinQuery = { __typename?: 'Query', coin?: { __typename?: 'Coin', id: any, owner: any, amount: number, color: any, maturity: number, status: CoinStatus, blockCreated: number } | null | undefined };
-
-export type DryRunMutationVariables = Exact<{
-  encodedTransaction: Scalars['HexString'];
-}>;
-
-
-export type DryRunMutation = { __typename?: 'Mutation', dryRun: Array<{ __typename?: 'Receipt', rawPayload: any }> };
-
-export type SubmitMutationVariables = Exact<{
-  encodedTransaction: Scalars['HexString'];
-}>;
-
-
-export type SubmitMutation = { __typename?: 'Mutation', submit: any };
-
-export type StartSessionMutationVariables = Exact<{ [key: string]: never; }>;
-
-
-export type StartSessionMutation = { __typename?: 'Mutation', startSession: string };
-
-export type EndSessionMutationVariables = Exact<{
-  sessionId: Scalars['ID'];
-}>;
-
-
-export type EndSessionMutation = { __typename?: 'Mutation', endSession: boolean };
-
-export type ExecuteMutationVariables = Exact<{
-  sessionId: Scalars['ID'];
-  op: Scalars['String'];
-}>;
-
-
-export type ExecuteMutation = { __typename?: 'Mutation', execute: boolean };
-
-export type ResetMutationVariables = Exact<{
-  sessionId: Scalars['ID'];
-}>;
-
-
-export type ResetMutation = { __typename?: 'Mutation', reset: boolean };
